+++ src/QuickStart/Program.cs
@@ -270,7 +270,7 @@ void WritePropertyComment()
     using var textReader = new StringReader(ini);
     var iniDocument = IniDocument.LoadFrom(textReader, IniDocumentConfiguration.Default);
 
-    iniDocument.WriteComment("Section", "Property", comment, CommentWriteMode.Fail);
+    iniDocument.WriteComment("Section", "Property", comment);
 
     using var textWriter = new StringWriter();
     iniDocument.SaveTo(textWriter);